split api tests into different files
