+++ .circleci/config.yml
@@ -268,7 +268,7 @@ jobs:
           environment:
             MOZ_HEADLESS: 1
             PYTEST_ADDOPTS: -n 1 --reruns 2
-          command: py -3.9 -m pytest tests\api\test_addon_uploads.py --driver Firefox --variables dev.json --html=api_submission_tests.html --self-contained-html
+          command: py -3.9 -m pytest tests\api --driver Firefox --variables dev.json --html=api_submission_tests.html --self-contained-html
           no_output_timeout: 30m
       - store_artifacts:
           path: api_submission_tests.html
